# docs/openapi/laravel-v1.yaml
openapi: 3.1.0
info:
  title: Learning English API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
paths:
...
  /api/v1/vocabulary:
    get:
      summary: Read locally persisted vocabulary
      operationId: listVocabulary
      security: []
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: per_page
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Vocabulary page
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Vocabulary" }
                  meta: { type: object }
...
  /api/v1/profile:
    put:
      summary: Update the current profile
      operationId: updateProfile
      security: [{ session: [], csrf: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "200": { $ref: "#/components/responses/UserResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "419": { $ref: "#/components/responses/CsrfError" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      summary: Delete the current account
      operationId: deleteProfile
      security: [{ session: [], csrf: [] }]
      parameters:
        - { $ref: "#/components/parameters/RequestId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "204": { description: Account deleted }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "409": { $ref: "#/components/responses/Conflict" }
        "419": { $ref: "#/components/responses/CsrfError" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /api/v1/progress:
    get:
      summary: Get personal course and lesson progress
      operationId: getLearnerProgress
      security: [{ session: [] }]
      responses:
        "200": { $ref: "#/components/responses/CollectionResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /api/v1/progress/dashboard:
    get:
      summary: Get the current learner's progress dashboard
      operationId: getLearnerProgressDashboard
      security: [{ session: [] }]
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
...
  /api/v1/admin/users:
    get:
      summary: List users without learner evidence
      operationId: listAdminUsers
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PerPage" }]
      responses:
        "200": { $ref: "#/components/responses/CollectionResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/users/{user}:
    put:
      summary: Update an ordinary user
      operationId: updateAdminUser
      security: [{ session: [], csrf: [] }]
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/RequestId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                role: { type: string, enum: [learner] }
              additionalProperties: false
      responses:
        "200": { $ref: "#/components/responses/DataResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "419": { $ref: "#/components/responses/CsrfError" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      summary: Anonymize and delete an ordinary user
      operationId: deleteAdminUser
      security: [{ session: [], csrf: [] }]
      parameters:
        - { $ref: "#/components/parameters/UserId" }
        - { $ref: "#/components/parameters/RequestId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 1 }
              additionalProperties: false
      responses:
        "204": { description: User deleted and retained records anonymized }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "419": { $ref: "#/components/responses/CsrfError" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  parameters:
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      in: query
      name: per_page
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    RequestId:
      in: header
      name: X-Request-ID
      required: true
      schema: { type: string, format: uuid }
      description: Idempotency key. Reuse with a different normalized payload returns 409.
...
    UserId:
      in: path
      name: user
      required: true
      schema: { type: integer, minimum: 1 }
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: english-learning-session
      description: Laravel web session cookie.
    csrf:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: URL-decoded XSRF-TOKEN cookie value required for mutations.
  schemas:
    User:
      type: object
      required: [id, name, email, email_verified_at, role]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        email_verified_at: { type: [string, "null"], format: date-time }
        role: { type: [string, "null"] }
    UserEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/User" }
        meta: { type: object }
    Message:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    ValidationError:
      type: object
      required: [message, errors]
      properties:
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    Vocabulary:
      type: object
      required: [id, word, meaning]
      properties:
        id: { type: integer }
        external_id: { type: [string, "null"] }
        word: { type: string }
        meaning: { type: string }
        definition: { type: [string, "null"] }
        translation: { type: [object, "null"] }
        pronunciation: { type: [string, "null"] }
        part_of_speech: { type: [string, "null"] }
        difficulty_level: { type: [string, "null"] }
        tags: { type: [array, "null"] }
        external_audio_url: { type: [string, "null"] }
    ResponseMeta:
      type: object
      properties:
        request_id: { type: ["string", "null"], format: uuid }
        generated_at: { type: string, format: date-time }
      additionalProperties: false
    ContractData:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum:
            - category
            - course
            - vocabulary
            - enrollment
            - learning_plan
            - learning_session
            - learning_attempt
            - lesson
            - learning_event
            - intervention_note
            - fsrs_card
            - fsrs_stats
            - assignment
            - progress
            - user
            - supervision_alert
            - sync_run
            - service_status
            - quota_policy
            - alert_rule
            - audit_event
            - teacher_assignment
            - speech_to_text
            - pronunciation
        id: { type: [integer, string, "null"] }
        name: { type: string }
        email: { type: string, format: email }
        role_id: { type: ["integer", "null"] }
        role: { type: [object, "null"] }
        title: { type: string }
        slug: { type: string }
        status: { type: string }
        user_id: { type: integer, minimum: 1 }
        learner_id: { type: integer, minimum: 1 }
        teacher_id: { type: integer, minimum: 1 }
        course_id: { type: integer, minimum: 1 }
        lesson_id: { type: ["integer", "null"], minimum: 1 }
        vocabulary_id: { type: ["integer", "null"], minimum: 1 }
        assignment_id: { type: ["integer", "null"], minimum: 1 }
        state: { type: string }
        due_at: { type: ["string", "null"], format: date-time }
        completed_at: { type: ["string", "null"], format: date-time }
        revision: { type: integer, minimum: 0 }
        due_count: { type: integer, minimum: 0 }
        new_count: { type: integer, minimum: 0 }
        learning_count: { type: integer, minimum: 0 }
        review_count: { type: integer, minimum: 0 }
        word: { type: string }
        meaning: { type: string }
        definition: { type: ["string", "null"] }
        example: { type: ["string", "null"] }
        pronunciation: { type: ["string", "null"] }
        audio_url: { type: ["string", "null"] }
        request_id: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 4 }
        step: { type: ["integer", "null"], minimum: 0 }
        stability: { type: ["number", "null"], minimum: 0 }
        difficulty: { type: ["number", "null"], minimum: 0 }
        scheduled_days: { type: ["integer", "null"], minimum: 0 }
        algorithm: { type: string }
        average_stability: { type: number, minimum: 0 }
        average_difficulty: { type: number, minimum: 0 }
        best_score: { type: ["integer", "null"], minimum: 0, maximum: 100 }
        severity: { type: string }
        rule_code: { type: string }
        service: { type: string }
        healthy: { type: boolean }
        version: { type: [integer, string] }
        action: { type: string }
        transcript: { type: string, maxLength: 2000 }
        pronunciation_score: { type: number, minimum: 0, maximum: 100 }
        estimated_minutes: { type: integer, minimum: 0 }
        event_id: { type: integer, minimum: 1 }
        is_correct: { type: boolean }
        summary: { type: [object, "null"] }
        activity: { type: [object, "null"] }
        progress: { type: object }
        course: { type: object }
        lesson: { type: [object, "null"] }
        progress_percent: { type: number, minimum: 0, maximum: 100 }
        completed_lessons: { type: integer, minimum: 0 }
        total_lessons: { type: integer, minimum: 0 }
        overview: { type: object }
        recent_activity: { type: array }
        recent_attempts: { type: array }
        features: { type: object }
        services: { type: object }
        open_alerts: { type: integer, minimum: 0 }
        trace_cag: { type: object }
        last_24_hours: { type: integer, minimum: 0 }
        last_30_days: { type: integer, minimum: 0 }
        degraded_30_days: { type: integer, minimum: 0 }
        limits: { type: object }
        is_active: { type: boolean }
        enabled: { type: boolean }
        parameters: { type: object }
        rule_key: { type: string }
        target_type: { type: ["string", "null"] }
        target_id: { type: ["string", "null"] }
        occurred_at: { type: ["string", "null"], format: date-time }
        evidence: { type: array }
        learner: { type: object }
        vocabulary: { type: [object, "null"] }
        instructions: { type: ["string", "null"] }
        resolution_note: { type: ["string", "null"] }
        items:
          type: array
          items:
            type: object
            required: [id, type, priority]
            properties:
              id: { type: [integer, string] }
              type: { type: string, enum: [teacher_lesson, teacher_vocabulary_practice, fsrs_review, course_activity, pronunciation_task, remediation] }
              priority: { type: integer, minimum: 1 }
            additionalProperties: false
      additionalProperties: false
    DataEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: "#/components/schemas/ContractData" }
        meta: { $ref: "#/components/schemas/ResponseMeta" }
      additionalProperties: false
    CollectionEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/ContractData" }
        meta:
          type: object
          required: [page, per_page, total]
          properties:
            page: { type: integer, minimum: 1 }
            per_page: { type: integer, minimum: 1, maximum: 100 }
            total: { type: integer, minimum: 0 }
            request_id: { type: ["string", "null"], format: uuid }
          additionalProperties: false
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required: [message, code]
      properties:
        message: { type: string }
        code: { type: string }
      additionalProperties: false
...
  responses:
    DataResponse:
      description: Stable data envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/DataEnvelope" }
    CollectionResponse:
      description: Paginated stable data envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CollectionEnvelope" }
...
    Forbidden:
      description: The authenticated role or assigned scope does not permit this action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: Idempotency key conflict, stale revision, or incompatible state transition
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    UserResponse:
      description: User payload
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UserEnvelope" }
...
    ValidationError:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
    Unauthenticated:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Message" }
    CsrfError:
      description: CSRF token mismatch
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Message" }